Improve UI of 404 page
